requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic-settings>=2.2.1
SQLAlchemy>=2.0.27
alembic>=1.13.1
asyncpg>=0.29.0
aioodbc>=0.5.0
redis>=5.0.1
httpx>=0.27.0
cryptography>=42.0.5
oracledb>=2.5.0
